Refactor to eliminate dylint dependency on clap
