+++ dylint/Cargo.toml
@@ -13,7 +13,6 @@ anyhow = "1.0.38"
 cargo = { version = "0.53.0", optional = true }
 cargo-platform = { version = "0.1.1", optional = true }
 cargo_metadata = { version = "0.13.1", optional = true }
-clap = "3.0.0-beta.2"
 glob = { version = "0.3.0", optional = true }
 if_chain = { version = "1.0.1", optional = true }
 lazy_static = "1.4.0"
